gifHistory

+++ src/app/users/pages/list-users/list-users.component.ts
@@ -3,28 +3,26 @@ import { UserService } from '../../services/user.service';
 import { User } from '../../interfaces/User.interface';
 import { CommonModule } from '@angular/common';
 import { firstValueFrom, Observable } from 'rxjs';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-list-users',
   standalone: true,
   imports: [CommonModule],
   templateUrl: './list-users.component.html',
-  styleUrl: './list-users.component.css'
+  styleUrl: './list-users.component.css',
 })
 export class ListUsersComponent implements OnInit {
+  public user: User[] = [];
 
-  public user:User[]=[]
+  constructor(private userService: UserService) {}
 
-  constructor(
-    private userService:UserService
-  ){}
-
-  get users(){
-    return [...this.userService.usersList]
+  get users() {
+    return [...this.userService.usersList];
   }
 
   ngOnInit(): void {
-    this.getUsers()
+    this.getUsers();
   }
 
   async getUsers(): Promise<void> {
@@ -36,9 +34,25 @@ export class ListUsersComponent implements OnInit {
       console.error('Error al obtener usuarios:', err);
     }
   }
-
-  deleteUserList(id:string):void{
-    const deletedUser=this.userService.deleteUserById(id)
-    this.user=deletedUser
+  deleteUserList(id: string): void {
+    Swal.fire({
+      title: 'Está seguro ?',
+      text: 'No podrá revertir esta acción',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, eliminar',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        const deletedUser = this.userService.deleteUserById(id);
+        this.user = deletedUser;
+        Swal.fire({
+          title: 'Eliminado!',
+          text: 'Usuario eliminado exitosamente!',
+          icon: 'success',
+        });
+      }
+    });
   }
 }
